requirements.txt
Flask
flask-cors
mtcnn
tensorflow
numpy
Pillow
opencv-python
keras-facenet
scipy